+++ packages/agent/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/agent [1.0.0-alpha.29](https://github.com/anolilab/lunora/compare/%40lunora%2Fagent%401.0.0-alpha.28...%40lunora%2Fagent%401.0.0-alpha.29) (2026-07-28)
+
+
+### Dependencies
+
+* **@lunora/ai:** upgraded to 1.0.0-alpha.42
+* **@lunora/mail:** upgraded to 1.0.0-alpha.35
+* **@lunora/server:** upgraded to 1.0.0-alpha.50
+
 ## @lunora/agent [1.0.0-alpha.28](https://github.com/anolilab/lunora/compare/%40lunora%2Fagent%401.0.0-alpha.27...%40lunora%2Fagent%401.0.0-alpha.28) (2026-07-27)