+++ packages/auth/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/auth [1.0.0-alpha.56](https://github.com/anolilab/lunora/compare/%40lunora%2Fauth%401.0.0-alpha.55...%40lunora%2Fauth%401.0.0-alpha.56) (2026-07-28)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.50
+
 ## @lunora/auth [1.0.0-alpha.55](https://github.com/anolilab/lunora/compare/%40lunora%2Fauth%401.0.0-alpha.54...%40lunora%2Fauth%401.0.0-alpha.55) (2026-07-27)