+++ packages/codegen/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/codegen [1.0.0-alpha.74](https://github.com/anolilab/lunora/compare/%40lunora%2Fcodegen%401.0.0-alpha.73...%40lunora%2Fcodegen%401.0.0-alpha.74) (2026-07-28)
+
+
+### Dependencies
+
+* **@lunora/advisor:** upgraded to 1.0.0-alpha.53
+* **@lunora/agent:** upgraded to 1.0.0-alpha.29
+* **@lunora/server:** upgraded to 1.0.0-alpha.50
+
 ## @lunora/codegen [1.0.0-alpha.73](https://github.com/anolilab/lunora/compare/%40lunora%2Fcodegen%401.0.0-alpha.72...%40lunora%2Fcodegen%401.0.0-alpha.73) (2026-07-27)