+++ packages/config/CHANGELOG.md
@@ -1,3 +1,12 @@
+## @lunora/config [1.0.0-alpha.100](https://github.com/anolilab/lunora/compare/%40lunora%2Fconfig%401.0.0-alpha.99...%40lunora%2Fconfig%401.0.0-alpha.100) (2026-07-28)
+
+
+### Dependencies
+
+* **@lunora/codegen:** upgraded to 1.0.0-alpha.74
+* **@lunora/seed:** upgraded to 1.0.0-alpha.49
+* **@lunora/studio:** upgraded to 1.0.0-alpha.78
+
 ## @lunora/config [1.0.0-alpha.99](https://github.com/anolilab/lunora/compare/%40lunora%2Fconfig%401.0.0-alpha.98...%40lunora%2Fconfig%401.0.0-alpha.99) (2026-07-27)