+++ packages/mail/CHANGELOG.md
@@ -1,3 +1,5 @@
+## @lunora/mail [1.0.0-alpha.35](https://github.com/anolilab/lunora/compare/%40lunora%2Fmail%401.0.0-alpha.34...%40lunora%2Fmail%401.0.0-alpha.35) (2026-07-28)
+
 ## @lunora/mail [1.0.0-alpha.34](https://github.com/anolilab/lunora/compare/%40lunora%2Fmail%401.0.0-alpha.33...%40lunora%2Fmail%401.0.0-alpha.34) (2026-07-27)
 
 ## @lunora/mail [1.0.0-alpha.33](https://github.com/anolilab/lunora/compare/%40lunora%2Fmail%401.0.0-alpha.32...%40lunora%2Fmail%401.0.0-alpha.33) (2026-07-27)