+++ packages/server/CHANGELOG.md
@@ -1,3 +1,5 @@
+## @lunora/server [1.0.0-alpha.50](https://github.com/anolilab/lunora/compare/%40lunora%2Fserver%401.0.0-alpha.49...%40lunora%2Fserver%401.0.0-alpha.50) (2026-07-28)
+
 ## @lunora/server [1.0.0-alpha.49](https://github.com/anolilab/lunora/compare/%40lunora%2Fserver%401.0.0-alpha.48...%40lunora%2Fserver%401.0.0-alpha.49) (2026-07-27)
 
 ## @lunora/server [1.0.0-alpha.48](https://github.com/anolilab/lunora/compare/%40lunora%2Fserver%401.0.0-alpha.47...%40lunora%2Fserver%401.0.0-alpha.48) (2026-07-27)