+++ packages/x402/CHANGELOG.md
@@ -1,3 +1,5 @@
+## @lunora/x402 [1.0.0-alpha.23](https://github.com/anolilab/lunora/compare/%40lunora%2Fx402%401.0.0-alpha.22...%40lunora%2Fx402%401.0.0-alpha.23) (2026-07-28)
+
 ## @lunora/x402 [1.0.0-alpha.22](https://github.com/anolilab/lunora/compare/%40lunora%2Fx402%401.0.0-alpha.21...%40lunora%2Fx402%401.0.0-alpha.22) (2026-07-27)
 
 ## @lunora/x402 [1.0.0-alpha.21](https://github.com/anolilab/lunora/compare/%40lunora%2Fx402%401.0.0-alpha.20...%40lunora%2Fx402%401.0.0-alpha.21) (2026-07-27)